Edited the rm_background method stub to change the interface and docstring.
* Added rm_ to the original method name. I generally like verbs in methods because methods do something and verbs describe action. Also, you should be aware of PEP 8, otherwise known as the Style Guide for Python Code. I added a section to this project's wiki that contains notes on my style conventions (https://github.com/jrsmith3/tfan-parsers/wiki/Style-Conventions) that you might want to look at. It has a link to PEP 8.
* Changed some of the names of the input arguments and added defaults for some. Here is a breakdown of what changed.
** loBE, hiBE: Initially you had variables for indices here. This requirement means that the user has to know something about how the data in the object is structured, and is frankly too much to ask of the user. Here is an example of why this construction is a problem: Let's say the user has two sets of data from the Si (2p) peaks between binding energies of 96.0eV and 106.0eV. The first set of data (dat1.dat) was collected with Delta E = 0.1eV and the second (dat2.dat) has Delta E = 0.2eV. Therefore, the first set of data has twice as many points as the second one. If the user is trying to analyze the size of the peak due to SiO2 at 103.7eV, the user will have to manually go through both sets of data in order to work out what indices to pass to the method. The user will find this scenario frustrating and time consuming.

A better approach is for the user just to specify the values of energy for the lower and upper bound of the interval. In this way the user does not have to know specifics of the data because the computer is responsible for dealing with the tedium of futzing with indices. It turns out that the algorithm for pulling out a sub-array of an array between a given interval is pretty easy. Even if the algorithm was complex, it is better to work out the algorithm up front and design an interface that is easy to use rather than leave the hard, tedious work up to the user. The programmer only has to solve the algorithm problem once and implement it in order to make the user's life better.

For both input arguments I'm explicitly setting the default to be the first and last elements of the binding energy array.

** size: It wasn't explicit what the size of the returned array was going to be. I'm guessing it would be how I specified it, but by adding this input argument the size question is made explicit. I am defaulting to 0 (a nonsense size for an array) because 0 can be interpreted as "default."

** model: I'm setting the default of the model to "linear." This might not be the right choice, but you don't want the method to fail because the user forgot to add this argument to the method call. Also, you want a default that is typical so that the user doesn't have to manually add an argument every time they call the function. Users are lazy and you should let them be lazy.

* Shortened the first line of the docstring.
* Edited the body of the docstring to give a motivation for the use of this method. Also gave an overview of how to use the method.
* Edited the list of input arguments to reflect the changes made to the interface. Added a set of allowed values for the "model" argument. Without this list, the user won't know what is legitimate input for this method.

The only thing this docstring is missing now are some doctests (http://docs.python.org/library/doctest.html). I've never actually written any doctests, but I think they are useful and a good idea.

One thing I'm not sure of yet is the "key" construction in the argument list. I know that the smooth and differentiate methods use this construction, but I've never been happy with it. Generally its a bad idea to change the API after you've tagged a revision of code for release, but I want to fix this clunky interface quirk if you have a good idea.